restructure runtime files
This commit shuffles the runtime files around due to platforms blurring
the lines between server and client. Now that the platform abstraction
exists, erlang can be used for client apps (for example, as a TUI app),
and so the previous deliniation is no longer relevant.

The renaming is not quite correct yet -- for example spa.ffi.ts was
renamed to a thin platform.ffi.mjs, but it may be redundant. Also
component.ffi.mjs and server_component.ffi.mjs are not exactly runtimes
yet they share the folder.
