+++ src/lustre/runtime/server_component.ffi.mjs
@@ -4,12 +4,12 @@
 // used as the entry module when running esbuild so we *cant* use imports relative
 // to src/.
 
-import { Reconciler } from "../../../../build/dev/javascript/lustre/lustre/vdom/reconciler.ffi.mjs";
+import { Reconciler } from "../../../build/dev/javascript/lustre/lustre/vdom/reconciler.ffi.mjs";
 import {
   adoptStylesheets,
   ContextRequestEvent,
-} from "../../../../build/dev/javascript/lustre/lustre/runtime/client/runtime.ffi.mjs";
-import { dom_strict } from "../../../../build/dev/javascript/lustre/lustre/platform.mjs";
+} from "../../../build/dev/javascript/lustre/lustre/runtime/platform/base.ffi.mjs";
+import { dom_strict } from "../../../build/dev/javascript/lustre/lustre/platform.mjs";
 import {
   mount_kind,
   reconcile_kind,
@@ -20,7 +20,7 @@ import {
   event_fired_kind,
   batch_kind,
   context_provided_kind,
-} from "../../../../build/dev/javascript/lustre/lustre/runtime/transport.mjs";
+} from "../../../build/dev/javascript/lustre/lustre/runtime/transport.mjs";
 
 //
